requirements.txt
streamlit
pandas
psycopg2-binary
xlsxwriter
openpyxl
python-docx
python-dotenv
reportlab
streamlit-autorefresh